Update diet_sensitivity_test.R
Two if loops (if metric == "trophic level") and (if metric == "biomass") were implemented but are almost identical. These loops are merged into one in this suggestion. Checking if the parameter metric is accurate is also done at the beginning instead of the end to abort right away if the wrong parameters are given. This version may reduce processing time.